% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_paired_trees.R
\name{add_scalebar}
\alias{add_scalebar}
\title{Add scale bar to cophylo plot}
\usage{
add_scalebar(obj, scale.bar, fsize)
}
\arguments{
\item{obj}{An object of class `cophy`}

\item{scale.bar}{A list of 2 numeric values to indicate position of scale bars on plot}

\item{fsize}{Font size of scale bar}
}
\description{
This function plots a host and symbiont tree given the object returned by
`sim_cophylo_bdp`.
}
\keyword{Internal}
