% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cophylo_summary_stat.R
\name{cophylo_summary_stat_by_indx}
\alias{cophylo_summary_stat_by_indx}
\alias{cophylo_summary_stat}
\title{Calculates summary statistics for cophylogenetyic objects}
\usage{
cophylo_summary_stat_by_indx(cophylo_obj, cophylo_obj_indx)

cophylo_summary_stat(cophylo_obj)
}
\arguments{
\item{cophylo_obj}{The cophylogenetic object produced via `sim_cophylo_bdp`}

\item{cophylo_obj_indx}{The index with `cophylo_obj` for `cophylo_summary_stat_by_indx`}
}
\value{
A vector consisting of (in order) cospeciations, host speciations, host extinctions, symbiont speciations, symbiont extinctions, parafit statistic, and parafit p-value

A dataframe containing statistics relevant to cophylogenetic analysis
}
\description{
For cophylogenetic objects produced in treeducken via `sim_cophylo_bdp`, calculates the numbers of different events of interest. In addition, calculates and tests the ParaFit test.
}
\section{Functions}{
\itemize{
\item \code{cophylo_summary_stat_by_indx}: Calculates the summary statistics for one index of the list of cophylogenetic objects
}}

\examples{

}
