% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cophy.R
\name{draw_curve}
\alias{draw_curve}
\title{Curve draw function}
\usage{
draw_curve(x, y, scale = 0.01, ...)
}
\arguments{
\item{x}{x positions on graph}

\item{y}{y positions on graph}

\item{scale}{Scale of the logistic (which is where the curve comes from)}

\item{...}{Other plotting parameters}
}
\description{
internal function to draw curved links between tips modified from Liam Revell phytools package under GPL v. 2
}
\references{
Revell, L.J. (2012), phytools: an R package for phylogenetic comparative biology (and other things). Methods in Ecology and Evolution, 3: 217-223. doi:10.1111/j.2041-210X.2011.00169.x
}
\author{
Wade Dismukes and Liam J Revell
}
\keyword{Internal}
