% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cophy.R
\name{get_ana_events}
\alias{get_ana_events}
\title{Internal function for plotting anagenetic events}
\usage{
get_ana_events(symb_tree, symb_coords, anagenetic_event_df)
}
\arguments{
\item{symb_tree}{the symb tree on which to plot the events}

\item{symb_coords}{the xy coordinates of `symb_tree` (obtained from the ape::plotPhyloCoor() function)}

\item{anagenetic_event_df}{a data frame formatted with these columns Host Index, Symbiont Index, Event Type, Event Time}
}
\description{
Internal function for plotting anagenetic events
}
\keyword{internal}
