% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.cophylo.R
\name{is.cophylo}
\alias{is.cophylo}
\alias{is.multiCophylo}
\title{Test for the cophylogenetic set object}
\usage{
is.cophylo(cophy)

is.multiCophylo(multiCophy)
}
\arguments{
\item{cophy}{an object to test to see if it is of class `cophylo`}

\item{multiCophy}{an object to test for multiCophy}
}
\value{
A logical vector
}
\description{
Tests if an object is of class `cophylo`
}
\details{
Checks that an object is of class `cophylo`. For multicophy checks that
the class is `multiCophylo` and that each element is of class `cophylo`.
}
\section{Functions}{
\itemize{
\item \code{is.multiCophylo}: Tests for `multiCophylo` composed of `cophylo` objects
}}

\examples{
h_lambda <- 1.0
h_mu <- 0.3
c_lambda <- 0.0
s_lambda <- 1.0
s_mu <- 0.3
s_her <- 0.0
host_symb_sets <- sim_cophylo_bdp(hbr = h_lambda,
                                  hdr = h_mu,
                                  sbr = s_lambda,
                                  cosp_rate = c_lambda,
                                  sdr = s_mu,
                                  host_exp_rate = s_her,
                                  timeToSimTo = 2.0,
                                  numbsim = 1)
is.cophylo(host_symb_sets[[1]])
is.multiCophylo(host_symb_sets)
}
\seealso{
as.cophylo
}
