% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_paired_trees.R
\name{make_links}
\alias{make_links}
\title{Internal tree plot function}
\usage{
make_links(
  obj,
  x,
  link.type = "curved",
  link.lwd = 1,
  link.col = "red",
  link.lty = "dashed"
)
}
\arguments{
\item{obj}{Object of class `cophy`}

\item{x}{x positions of beginning and end of links}

\item{link.type}{Type of link "curved" or "straight"}

\item{link.lwd}{Link line width}

\item{link.col}{link color}

\item{link.lty}{Link style "dashed" being default}
}
\description{
internal function to make links between tips modified from Liam Revell phytools package under GPL v. 2
}
\details{
This is the most different from phytools implementation in order to allow for extinct tips in the the cophy plot. In addition to involve less wrangling for the cophy object.
}
\references{
Revell, L.J. (2012), phytools: an R package for phylogenetic comparative biology (and other things). Methods in Ecology and Evolution, 3: 217-223. doi:10.1111/j.2041-210X.2011.00169.x
}
\keyword{Internal}
