% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cophy.R
\name{make_textbox}
\alias{make_textbox}
\title{Internal tree plot function}
\usage{
make_textbox(x, y, label, pos, offset, cex, font)
}
\arguments{
\item{x}{x coordinates}

\item{y}{y coordinates}

\item{label}{Labels as vector of strings}

\item{pos}{Position in plot environment}

\item{offset}{How offset from tips}

\item{cex}{a numerical vector giving the amount by which characters
should be scaled relative to the default}

\item{font}{font choice}
}
\description{
internal function to make textbox for tip labels
modified from phytools::TEXTBOX package under GPL v. 2
}
\references{
Revell, L.J. (2012), phytools: an R package for phylogenetic comparative
biology (and other things). Methods in Ecology and Evolution, 3: 217-223.
doi:10.1111/j.2041-210X.2011.00169.x
}
\author{
Wade Dismukes and Liam J Revell
}
\keyword{Internal}
