% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_paired_trees.R
\name{phylogram}
\alias{phylogram}
\title{Internal tree plot function}
\usage{
phylogram(
  tree,
  part = 1,
  direction = "rightwards",
  fsize = 1,
  ftype = "i",
  lwd = 1,
  ...
)
}
\arguments{
\item{tree}{an object of class `phylo`}

\item{part}{Part of page to plot on}

\item{direction}{Leftwards or rightwards facing phylogeny on plot}

\item{fsize}{Font size of labels on plot}

\item{ftype}{What type of font to use}

\item{lwd}{Plot line width}

\item{...}{Other plotting parameters}
}
\description{
internal function to plot modified from Liam Revell phytools package under GPL v. 2
}
\references{
Revell, L.J. (2012), phytools: an R package for phylogenetic comparative biology (and other things). Methods in Ecology and Evolution, 3: 217-223. doi:10.1111/j.2041-210X.2011.00169.x
}
\keyword{Internal}
