% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cophy.R
\name{plot.cophy}
\alias{plot.cophy}
\alias{plot.multiCophy}
\title{Plot host and symbiont pair with current associations}
\usage{
\method{plot}{cophy}(
  x,
  use_edge_length = TRUE,
  type = "phylogram",
  col = par("fg"),
  lwd = par("lwd"),
  lty = par("lty"),
  show_tip_label = TRUE,
  gap = 1,
  font = 3,
  fsize = 1,
  show_scalebar = FALSE,
  scalebar_fsize = 0,
  ...
)

\method{plot}{multiCophy}(x, ...)
}
\arguments{
\item{x}{object of class multiCophy}

\item{use_edge_length}{Boolean to draw trees with edge length or not}

\item{type}{string "phylogram" or "cladogram"}

\item{col}{What color to draw links between trees}

\item{lwd}{Width of links between trees}

\item{lty}{Type of line to draw between trees}

\item{show_tip_label}{Boolean for showing labels}

\item{gap}{Size of the gap between the tips and tip names}

\item{font}{What font to use (bold, italic (default), etc.)}

\item{fsize}{What size font as a character expansion factor (same as cex)}

\item{show_scalebar}{Show a scalebar}

\item{scalebar_fsize}{Font size of scalebars (default is 0 -> no scalebar)}

\item{...}{other plotting parameters}
}
\value{
a plot of the host and symbiont tree with extant interactions
}
\description{
This function plots a host and symbiont tree given the object returned by
`sim_cophylo_bdp`.
}
\details{
This function is mostly an altered version of the cophyloplot function
written by Damien de Vienne Copyright 2008 - 2010 under GPL.
}
\section{Functions}{
\itemize{
\item \code{plot.multiCophy}: Plots multiple cophy plots
}}

\examples{

tr_pair <- sim_cophylo_bdp(hbr = 0.1,
                           hdr = 0.05,
                           sdr = 0.1,
                           host_exp_rate = 0.4,
                           sbr = 0.05,
                           cosp_rate =1.0,
                           numbsim = 10,
                           time_to_sim = 2)
plot.cophy(tr_pair[[1]])
}
