% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_paired_trees.R
\name{plot.cophylo}
\alias{plot.cophylo}
\title{Plot host and symbiont pair with current associations}
\usage{
\method{plot}{cophylo}(x, ...)
}
\arguments{
\item{x}{a tree pair object returned by `sim_cophylo_bdp`}

\item{...}{other plotting parameters}
}
\value{
a plot of the host and symbiont tree with extant interactions
}
\description{
This function plots a host and symbiont tree given the object returned by
`sim_cophylo_bdp`.
}
\examples{

tr_pair <- sim_cophylo_bdp(hbr = 0.1,
                           hdr = 0.05,
                           sdr = 0.1,
                           host_exp_rate = 0.4,
                           sbr = 0.05,
                           cosp_rate =1.0,
                           numbsim = 10,
                           time_to_sim = 2)
plot.cophylo(tr_pair[[1]])
}
