% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_cophylo.R
\name{host_tree}
\alias{host_tree}
\alias{host_tree.cophylo}
\alias{symb_tree}
\alias{symb_tree.cophylo}
\alias{association_mat}
\alias{association_mat.cophylo}
\alias{print.cophylo}
\title{Print a cophylogenetic set}
\usage{
host_tree(cophy)

\method{host_tree}{cophylo}(cophy)

symb_tree(cophy)

\method{symb_tree}{cophylo}(cophy)

association_mat(cophy)

\method{association_mat}{cophylo}(cophy)

\method{print}{cophylo}(x, printlen = 24, ...)
}
\arguments{
\item{cophy}{An object of class `cophylo`}

\item{...}{Further arguments used in generic classes}

\item{object}{An object of class `cophylo` or class `multiCophylo`}
}
\value{
Print returns NULL. host_tree returns NULL, symb_tree returns NULL,
association_mat returns the dimensions of the matrix, event_history returns NULL.
}
\description{
Prints a cophylogenetic set or a list of cophylogenetic sets.
}
\section{Functions}{
\itemize{
\item \code{host_tree}: Returns host tree of a cophylogenetic set

\item \code{host_tree.cophylo}: Returns host tree  of a cophylogenetic set

\item \code{symb_tree}: Returns symb tree of a cophylogenetic set

\item \code{symb_tree.cophylo}: Returns symb tree of a cophylogenetic set

\item \code{association_mat}: Returns association matrix of a cophylogenetic set

\item \code{association_mat.cophylo}: Returns association matrix of a cophylogenetic set
}}

\examples{
h_lambda <- 1.0
h_mu <- 0.3
c_lambda <- 0.0
s_lambda <- 1.0
s_mu <- 0.3
s_her <- 0.0
host_symb_sets <- sim_cophylo_bdp(hbr = h_lambda,
                                  hdr = h_mu,
                                  sbr = s_lambda,
                                  cosp_rate = c_lambda,
                                  sdr = s_mu,
                                  host_exp_rate = s_her,
                                  timeToSimTo = 2.0,
                                  numbsim = 4)
print(host_symb_sets[[1]])
host_tree(host_symb_sets[[1]])
symb_tree(host_symb_sets[[1]])
association_mat(host_symb_sets[[1]])
event_history(host_symb_sets[[1]])
print(host_symb_sets)
}
\seealso{
`sim_cophylo_bdp`, `print` for the generic, `multiCophylo`, `c.cophylo`
}
