% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_cophylo.R
\name{host_tree.multiCophylo}
\alias{host_tree.multiCophylo}
\alias{symb_tree.multiCophylo}
\alias{association_mat.multiCophylo}
\alias{event_history.multiCophylo}
\alias{print.multiCophylo}
\title{describeIn print.cophylo Prints a list of cophylogenetic sets}
\usage{
\method{host_tree}{multiCophylo}(cophy)

\method{symb_tree}{multiCophylo}(cophy)

\method{association_mat}{multiCophylo}(cophy)

\method{event_history}{multiCophylo}(cophy)

\method{print}{multiCophylo}(x, details = FALSE, ...)
}
\arguments{
\item{cophy}{Object of class `multiCophylo`}

\item{details}{A logical value, outputs brief summary of each set in the list.}
}
\description{
describeIn print.cophylo Prints a list of cophylogenetic sets
}
\section{Functions}{
\itemize{
\item \code{host_tree.multiCophylo}: Returns host tree of each member of a list of cophylogenetic sets

\item \code{symb_tree.multiCophylo}: Returns symb tree of each member of a list of cophylogenetic sets

\item \code{association_mat.multiCophylo}: Returns association matrix for each member of a list of cophylogenetic sets

\item \code{event_history.multiCophylo}: Returns event_history for each member of a list of cophylogenetic sets
}}

