% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sim_locustree_bdp}
\alias{sim_locustree_bdp}
\title{Simulates locus tree using constant rate birth-death-transfer process}
\usage{
sim_locustree_bdp(species_tree_, gbr_, gdr_, lgtr_, num_loci_)
}
\arguments{
\item{species_tree_}{species tree to simulate along}

\item{gbr_}{gene birth rate}

\item{gdr_}{gene death rate}

\item{lgtr_}{gene trasnfer rate}

\item{num_loci_}{number of locus trees to simulate}
}
\value{
List of objects of the tree class (as implemented in APE)
}
\description{
Simulates locus tree using constant rate birth-death-transfer process
}
\details{
Given a species tree simulates a locus or gene family tree along
    the species tree.
}
\examples{
# first simulate a species tree
mu <- 0.5 # death rate
lambda <- 2.0 # birth rate
numb_replicates <- 10
numb_extant_tips <- 4
# simulate trees under the GSA so first simulates a tree with
# numb_extant_tips * 100 tips counting each time we have a tree with 10 tips
# then randomly picks one of those trees

sp_tree <- sim_sptree_bdp(sbr_ = lambda,
                sdr_ = mu,
                numbsim_ = numb_replicates,
                n_tips_ = numb_extant_tips)

gene_br <- 1.0
gene_dr <- 0.2
transfer_rate <- 0.2
sim_locustree_bdp(species_tree_ = sp_tree,
                  gbr_ = gene_br,
                  gdr_ = gene_dr,
                  lgtr_ = transfer_rate,
                  num_loci_ = 10)
}
\references{
Rasmussen MD, Kellis M. Unified modeling of gene duplication, loss, and
    coalescence using a locus tree. Genome Res. 2012;22(4):755–765.
    doi:10.1101/gr.123901.111
}
