% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sim_locustree_genetree_mlc}
\alias{sim_locustree_genetree_mlc}
\title{Simulate locus tree within species tree and gene trees within locus tree}
\usage{
sim_locustree_genetree_mlc(
  species_tree,
  gbr,
  gdr,
  lgtr,
  num_loci,
  num_sampled_individuals,
  theta,
  num_genes_per_locus
)
}
\arguments{
\item{species_tree}{input species tree of class "phylo"}

\item{gbr}{gene birth rate}

\item{gdr}{gene death rate}

\item{lgtr}{lateral gene transfer rate'}

\item{num_loci}{number of loci to simulate}

\item{num_sampled_individuals}{number of individuals sampled within each locus lineage}

\item{theta}{the population genetic parameter}

\item{num_genes_per_locus}{number of genes to simulate within each locus}
}
\value{
A list of lists of length 2. The first element of each list of length 2 is `locus.tree` the locus tree and the second element is a list of the gene trees simulated within that locus tree. All trees are of class "phylo".
}
\description{
First simulates a locus tree within the confines of the input species tree using a constant-rate birth-death process
based on values of `gbr`, `gdr` and `lgtr`. Then simulates gene trees within that locus tree using the multi-locus coalescent process.
}
\examples{
# first simulate a species tree
mu <- 0.5
lambda <- 1.0
nt <- 6
tr <- sim_sptree_bdp(sbr = lambda, sdr = mu, numbsim = 1, n_tips = nt)
# for a locus tree with 100 genes sampled per locus tree
loctr_gentr <- sim_locustree_genetree_mlc(tr[[1]],
                                           gbr = 0.1,
                                           gdr = 0.0,
                                           lgtr = 0.0,
                                           theta = 1,
                                           num_sampled_individuals = 1,
                                           num_loci = 4,
                                           num_genes_per_locus = 100)

}
\references{
Mallo D, de Oliveira Martins L, Posada D (2015) SimPhy: Phylogenomic Simulation of Gene, Locus and Species Trees. Syst. Biol. doi: http://dx.doi.org/10.1093/sysbio/syv082
}
\seealso{
sim_locustree_bdp
}
