% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sim_multispecies_coal}
\alias{sim_multispecies_coal}
\title{Simulate multispecies coalescent on a species tree}
\usage{
sim_multispecies_coal(
  species_tree,
  ne,
  num_sampled_individuals,
  num_genes,
  rescale = FALSE,
  mutation_rate = 1,
  generation_time = 1
)
}
\arguments{
\item{species_tree}{input species tree of class "phylo"}

\item{ne}{Effective population size}

\item{num_sampled_individuals}{number of individuals sampled within each lineage}

\item{num_genes}{number of genes to simulate within each locus}

\item{rescale}{Rescale the tree into coalescent units (otherwise assumes it is in those units)}

\item{mutation_rate}{The rate of mutation per generation}

\item{generation_time}{Number of generations per species tree length}
}
\value{
A list of coalescent trees
}
\description{
Simulates the multispecies coalescent on a species tree.
}
\examples{
# first simulate a species tree
mu <- 0.5
lambda <- 1.0
nt <- 6
tr <- sim_sptree_bdp(sbr = lambda, sdr = mu, numbsim = 1, n_tips = nt)
# for a locus tree with 100 genes sampled per locus tree
loctr_gentr <- sim_multispecies_coal(tr[[1]],
                                    ne = 1,
                                    num_sampled_individuals = 1,
                                    num_genes = 100)

}
\references{
Mallo D, de Oliveira Martins L, Posada D (2015) SimPhy: Phylogenomic Simulation of Gene, Locus and Species Trees. Syst. Biol. doi: http://dx.doi.org/10.1093/sysbio/syv082
}
\seealso{
sim_locustree_bdp, sim_sptree_bdp, sim_sptree_bdp_time
}
