% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_cophylo.R
\name{str.multiCophy}
\alias{str.multiCophy}
\title{Retrieve the structure of a class multiCophy}
\usage{
\method{str}{multiCophy}(object, ...)
}
\arguments{
\item{object}{An object of class multiCophy}

\item{...}{Potential further arguments to generic str class}
}
\value{
NULL value
}
\description{
Retrieves the structure of the class multiCophy
}
\examples{
h_lambda <- 1.0
h_mu <- 0.3
c_lambda <- 0.0
s_lambda <- 1.0
s_mu <- 0.3
s_her <- 0.0
host_symb_sets <-  sim_cophylo_bdp(hbr = h_lambda,
                                  hdr = h_mu,
                                  sbr = s_lambda,
                                  cosp_rate = c_lambda,
                                  sdr = s_mu,
                                  host_exp_rate = s_her,
                                  time_to_sim = 1.0,
                                  numbsim = 2)
str(host_symb_sets)

}
