\name{treeducken-package}
\alias{treeducken-package}
\alias{treeducken}
\docType{package}
\title{
treeducken: simulates cophylogenetic systems & nested phylogenies
}
\description{
  treeducken simulates cophylogenetic systems such as host and symbiont pairs.
  This is done using the \code{sim_cophylo_bdp} function. This function simulates
  a host tree and a symbiont tree simultaneously using a matrix describing the
  associations between contemporaneous hosts and symbionts. These simulations
  allow for varying rates of host-shift speciation and cospeciation in addition
  to independent birth and death rates of the host and the symbiont trees.
  treeducken is also able to simulate nested phylogenies such as might be
  expected in the case of gene tree and species tree scenarios.
}
\details{
\tabular{ll}{
Package: \tab treeducken\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2020-03-25\cr
License: \tab GPL-2  \cr
LazyLoad: \tab yes\cr
}
}
\author{
Wade Dismukes
<wade.dismukes@gmail.com>
}
\references{
  Dismukes W. and Tracy A. Heath, Something something (work-in-progress)
}
\keyword{ package }
\seealso{
\code{ape},
\code{geiger},
\code{TreeSim},
\code{PhyTools}
}
