% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cophy.R
\name{add_scalebar}
\alias{add_scalebar}
\title{Add scale bar to cophylo plot}
\usage{
add_scalebar(host_coords, symb_coords, fsize)
}
\arguments{
\item{host_coords}{Host x,y coordinates}

\item{symb_coords}{Symb x,y coordinates}

\item{fsize}{Font size of scale bar}
}
\description{
This function plots a host and symbiont tree given the object returned by
`sim_cophylo_bdp`.
}
\keyword{Internal}
