% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cophylo_summary_stat.R
\name{cophylo_summary_stat_by_indx}
\alias{cophylo_summary_stat_by_indx}
\alias{cophylo_summary_stat}
\title{Calculates summary statistics for cophylogenetyic objects}
\usage{
cophylo_summary_stat_by_indx(cophylo_obj, cophylo_obj_indx)

cophylo_summary_stat(cophylo_obj)
}
\arguments{
\item{cophylo_obj}{The cophylogenetic object produced via `sim_cophylo_bdp`}

\item{cophylo_obj_indx}{The index with `cophylo_obj` for `cophylo_summary_stat_by_indx`}
}
\value{
A vector consisting of (in order) cospeciations, host speciations, host extinctions, symbiont speciations, symbiont extinctions, parafit statistic, and parafit p-value

A dataframe containing statistics relevant to cophylogenetic analysis
}
\description{
For cophylogenetic objects produced in treeducken via `sim_cophylo_bdp`, calculates the numbers of different events of interest. In addition, calculates and tests the ParaFit test.
}
\section{Functions}{
\itemize{
\item \code{cophylo_summary_stat_by_indx}: Calculates the summary statistics for one index of the list of cophylogenetic objects
}}

\examples{
host_mu <- 0.5 # death rate
host_lambda <- 2.0 # birth rate
numb_replicates <- 10
time <- 1.0
symb_mu <- 0.2
symb_lambda <- 0.4
host_shift_rate <- 0.1
cosp_rate <- 2.0

cophylo_pair <- sim_cophylo_bdp(hbr = host_lambda,
                           hdr = host_mu,
                           cosp_rate = cosp_rate,
                           host_exp_rate = host_shift_rate,
                           sdr = symb_mu,
                           sbr = symb_lambda,
                           numbsim = numb_replicates,
                           time_to_sim = time)
summary_stats <- cophylo_summary_stat(cophylo_pair)
}
